requirements.txt
aiohttp==3.6.2
apiron==5.0.0
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.2
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.9
multidict==4.7.5
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
slackclient==2.5.0
soupsieve==2.0
urllib3==1.25.8
yarl==1.4.2
